requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
Babel==2.15.0
backoff==2.2.1
bangla==0.0.2
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.34.160
botocore==1.34.160
build==1.2.1
cachelib==0.13.0
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.18.1
cohere==5.8.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
crewai==0.51.1
crewai-tools==0.8.3
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dataclasses-json==0.6.7
datasets==2.20.0
dateparser==1.1.8
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docstring_parser==0.16
docx2txt==0.8
einops==0.8.0
embedchain==0.1.120
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.112.0
fastavro==1.9.5
filelock==3.15.4
Flask==3.0.3
Flask-Session==0.8.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
g2pkk==0.1.2
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-aiplatform==1.62.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
gptcache==0.1.44
greenlet==3.0.3
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.3
grpcio-tools==1.62.3
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
h2==4.1.0
hangul-romanize==0.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
iniconfig==2.0.0
instructor==1.3.3
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.25.3
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.2.13
langchain-cohere==0.1.9
langchain-community==0.2.12
langchain-core==0.2.30
langchain-experimental==0.0.64
langchain-groq==0.1.9
langchain-huggingface==0.0.3
langchain-openai==0.1.21
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.99
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
Mako==1.3.5
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.0.9
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
nodeenv==1.9.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.40.6
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pooch==1.8.2
portalocker==2.10.1
posthog==3.5.0
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylance==0.9.18
pynndescent==0.5.13
pyOpenSSL==24.2.1
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
PyPika==0.48.9
pypinyin==0.51.0
pyproject_hooks==1.1.0
pyright==1.1.375
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.2
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
qdrant-client==1.11.0
ratelimiter==1.2.0.post0
redis==5.0.8
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.4
schema==0.7.7
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.23.1
semver==3.0.2
sentence-transformers==3.0.1
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.4.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
tornado==6.4.1
tqdm==4.66.5
trainer==0.0.36
traitlets==5.14.3
transformers==4.44.0
trio==0.26.2
trio-websocket==0.11.1
triton==3.0.0
TTS==0.22.0
typeguard==4.3.0
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.23.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.0
